+++ firmware/hw_layer/mass_storage/mass_storage.mk
@@ -8,3 +8,4 @@ ALLINC += $(PROJECT_DIR)/ext/uzlib/src
 ALLCPPSRC += $(PROJECT_DIR)/hw_layer/mass_storage/null_device.cpp \
 			 $(PROJECT_DIR)/hw_layer/mass_storage/compressed_block_device.cpp \
 			 $(PROJECT_DIR)/hw_layer/mass_storage/mass_storage_device.cpp \
+			 $(PROJECT_DIR)/hw_layer/mass_storage/mass_storage_init.cpp \
