+++ firmware/hw_layer/mass_storage/mass_storage_init.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "hal.h"
+
+#if HAL_USE_USB_MSD
+void initUsbMsd();
+void attachMsdSdCard(BaseBlockDevice* blkdev);
+#endif
